% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_accountability.R
\name{plot_accountability}
\alias{plot_accountability}
\title{Plot Accountability Bars}
\usage{
plot_accountability(tracker_obj)
}
\arguments{
\item{tracker_obj}{An object of class \code{sustain_tracker}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Generates a horizontal stacked bar chart showing the proportion of status levels
per owner, ranked by the percentage of "Validated" data.
}
\examples{
\dontrun{
tracker <- sustain_ingest(sustain_dummy_data)
plot_accountability(tracker)
}
}
