% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_collection_trend.R
\name{plot_collection_trend}
\alias{plot_collection_trend}
\title{Plot Collection Trend}
\usage{
plot_collection_trend(tracker_obj)
}
\arguments{
\item{tracker_obj}{An object of class \code{sustain_tracker}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Generates a cumulative Burn-up chart of "Validated" items over time.
}
\examples{
\dontrun{
tracker <- sustain_ingest(sustain_dummy_data)
plot_collection_trend(tracker)
}
}
