% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quadrant.R
\name{plot_quadrant}
\alias{plot_quadrant}
\title{Plot Quadrant Chart}
\usage{
plot_quadrant(tracker_obj)
}
\arguments{
\item{tracker_obj}{An object of class \code{sustain_tracker}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Generates a scatter plot mapping priority against status.
Highlights data points that are "High" priority and "Not Delivered".
}
\examples{
\dontrun{
tracker <- sustain_ingest(sustain_dummy_data)
plot_quadrant(tracker)
}
}
