% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme_heatmap.R
\name{plot_theme_heatmap}
\alias{plot_theme_heatmap}
\title{Plot Thematic Heatmap}
\usage{
plot_theme_heatmap(tracker_obj)
}
\arguments{
\item{tracker_obj}{An object of class \code{sustain_tracker}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Generates a faceted tile plot representing individual data points within themes,
forming a "Data Grid" of maturity.
}
\examples{
\dontrun{
tracker <- sustain_ingest(sustain_dummy_data)
plot_theme_heatmap(tracker)
}
}
