% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_theme_network.R
\name{plot_theme_network}
\alias{plot_theme_network}
\title{Plot Interdependency Network Graph}
\usage{
plot_theme_network(tracker_obj, interactive = FALSE)
}
\arguments{
\item{tracker_obj}{An object of class \code{sustain_tracker}.}

\item{interactive}{Logical. If \code{TRUE}, returns a \code{visNetwork} object. If \code{FALSE} (default), returns a \code{ggraph} object.}
}
\value{
A \code{ggraph} or \code{visNetwork} object representing the data-to-theme dependencies.
}
\description{
Generates a network graph linking data nodes to their central theme nodes.
}
\examples{
\dontrun{
tracker <- sustain_ingest(sustain_dummy_data)
plot_theme_network(tracker, interactive = FALSE)
}
}
