% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_waffle_progress.R
\name{plot_waffle_progress}
\alias{plot_waffle_progress}
\title{Plot Waffle Chart of Progress}
\usage{
plot_waffle_progress(tracker_obj)
}
\arguments{
\item{tracker_obj}{An object of class \code{sustain_tracker}.}
}
\value{
A \code{ggplot2} object containing the waffle chart.
}
\description{
Generates a waffle chart grid representing the proportion of data collected
across the four status levels.
}
\examples{
\dontrun{
tracker <- sustain_ingest(sustain_dummy_data)
plot_waffle_progress(tracker)
}
}
